fix: return image if content-type cannot be infer
This is very particular to TracktikCloud S3 buckets objects. We have
uploaded a bunch of them with a generic content type
application/octet-stream and some valid images don't pass the validation
to infer the content type using the file header. Let's return by default
image as the content type instead of returning a 500 error until we
apply the fix to our objects.